fft out is not int8_t

+++ fft.hpp
@@ -16,12 +16,12 @@ public:
 
 	void configure(size_t size, Window::Type type, float beta=5.0f);
 	int out_size();
-	std::vector<float> run(Sample *input, size_t stride=1);
+	std::vector<int8_t> run(Sample *input, size_t stride=1);
 
 private:
 	fftwf_plan m_plan{nullptr};
 	Window m_window{};
 	std::vector<float> m_in;
-	std::vector<float> m_out;
-	std::map<double, std::vector<float>> m_cache;
+	std::vector<int8_t> m_out;
+	std::map<double, std::vector<int8_t>> m_cache;
 };
